Fix the logic of saving data to db
There was an issue with scraping the data and saving
movies and actors based on names since multiple movies or
actors can have same name.
In the scope of this commit, the script was changed to save
the id of the movie and actor and save it as a primary key thus
fixing the issue.